Fixed newsletter subscription bugs, removed debug puts
The problem I was having (subscriptions not appearing on Mailchimp site)
turned out to be because I was using example.com addresses, which
Mailchimp ignores. So this was fixed by simply using a more real-looking
test address.

There was also a problem with filling in the merge field for "name", but
again, this was fixed by fixing up the name of the field on the
Mailchimp end; nothing was needed here.

So in the end, all I really did was remove the puts statements.